Corrected some problems with the implementation of the update methods. The application is currently completely functional, but I want to clean up the properties "module" before merging this branch to main.

+++ BX.Business/Managers/Property/PropertyManager.cs
@@ -25,9 +25,9 @@ namespace BX.Business.Managers
             return await _propertyRepository.GetPropertiesAsync();
         }
 
-        public async Task<Property> GetPropertyAsync(int id)
+        public async Task<Property> GetPropertyByIdAsync(int id)
         {
-            return await _propertyRepository.GetPropertyAsync(id);
+            return await _propertyRepository.GetPropertyByIdAsync(id);
         }
 
         public async Task<IEnumerable<Property>> GetFilteredPropertiesAsync(string searchText, string status)
@@ -44,13 +44,6 @@ namespace BX.Business.Managers
         }
         #endregion
 
-        #region PUT
-        public async Task<bool> UpdatePropertyAsync(Property property)
-        {
-            return await _propertyRepository.UpdatePropertyAsync(property);
-        }
-        #endregion
-
         #region PATCH
         public async Task<bool> UpdatePropertyStatusAsync(Property property)
         {
@@ -58,7 +51,7 @@ namespace BX.Business.Managers
         }
         public async Task<bool> UpdatePropertyDetailsAsync(Property property)
         {
-            Property existingProperty = await _propertyRepository.GetPropertyAsync(property.Id);
+            Property existingProperty = await _propertyRepository.GetPropertyByIdAsync(property.Id);
             if (existingProperty == null)
             {
                 throw new Exception("Property not found");
@@ -82,14 +75,46 @@ namespace BX.Business.Managers
             existingProperty.IsCondominium = property.IsCondominium;
             return await _propertyRepository.UpdatePropertyDetailsAsync(property);
         }
+        public async Task<bool> UpdatePropertyAsync(Property newProperty)
+        {
+            Property existingProperty = await _propertyRepository.GetPropertyByIdAsync(newProperty.Id);
+            if (existingProperty==null) return false;
+            UpdatePropertyDetails(existingProperty, newProperty);
+            return await _propertyRepository.UpdatePropertyAsync(existingProperty);
+        }
         #endregion
 
         #region DELETE
-        public async Task<bool> DeletePropertyAsync(Property property)
+        public async Task<bool> DeletePropertyAsync(int id)
         {
+            Property property = await _propertyRepository.GetPropertyByIdAsync(id);
+            if (property==null)
+            {
+                return false;
+            }
             return await _propertyRepository.DeletePropertyAsync(property);
         }
         #endregion
 
+        private void UpdatePropertyDetails(Property existingProperty, Property property)
+        {
+            existingProperty.Name = property.Name;
+            existingProperty.Description = property.Description;
+            existingProperty.Price = property.Price;
+            existingProperty.Status = property.Status;
+            existingProperty.CreatedDate = property.CreatedDate;
+            existingProperty.ConstructionType = property.ConstructionType;
+            existingProperty.Province = property.Province;
+            existingProperty.Canton = property.Canton;
+            existingProperty.ConstructionSize = property.ConstructionSize;
+            existingProperty.LandSize = property.LandSize;
+            existingProperty.Bedrooms = property.Bedrooms;
+            existingProperty.Bathrooms = property.Bathrooms;
+            existingProperty.Price = property.Price;
+            existingProperty.Floors = property.Floors;
+            existingProperty.HasGarage = property.HasGarage;
+            existingProperty.GarageCapacity = property.GarageCapacity;
+            existingProperty.IsCondominium = property.IsCondominium;
+        }
     }
 }
